requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
onnxruntime>=1.15.0
requests
mtcnn-runtime
tqdm
starlette
python-multipart
